Remove redundant and misleading data
The external id field being unique for every entry meant we were
creating a new timeseries for every datapoint which was not intended.
The additional metadata also did not function as expected, given that
it was not tied to a specific datapoint and as such the specific
coordinates were confusing and not attached to the correct entity.